Move setup tooltip to GA4 controls.
